==> 2.Fit_3_models_example.Rmd <==
---
title: "2. Fit 3 lactation models"
output: 
  html_notebook: 
    fig_width: 8
    toc: yes
    code_folding: show
author: "David Innes"
editor_options: 
  chunk_output_type: inline
---

This step is slow, but is designed so that the `2.fit_3_lac_models_parallel.R` could be executed on a server for faster processing. 
We run it below to show terminal output, but other outputs are saved to `./Outputs/` folder and `/Plots` folder automatically.


Output: saves 4x dataframes:

1. Combined_params
2. Combined_fit_stats
3. Combined_preds
4. Combined fit stats
Also saves plots. 

It also leaves various objects in environment which can be perused as well, particularly if you want to access the actual fit objects. See the source code for how it is structured (in nested data frames).

It requires the output from step 1 to be saved as: `/Outputs/cleaned_data_out.rds`

# Requirements:
```{r eval=FALSE}
# these are not executed here, but are required by the script to run
suppressPackageStartupMessages(library(tidyverse, warn.conflicts = FALSE))
library(nlme)
library(nls.multstart)
library(broom)
library(purrr)
library(multidplyr)
library(car)


source("./functions/f_lac_equations.R")
source("./functions/f_goodness_of_fit.R")

```

# Execute:
This can be slow! Consider running it via command line on a server, if available. 
```{r}
source("2.fit_3_lac_models_parallel.R")
```

Example Output:
```{}
[1] "START"
[1] "2023-03-09 11:13:44 EST"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 11:14:39 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 11:20:38 EST"
[1] "Fitting WOOD..."
[1] "2023-03-09 11:20:38 EST"
[1] "Fitting WOOD...END"
[1] "2023-03-09 11:23:12 EST"
[1] "Fitting WILMINK..."
[1] "2023-03-09 11:23:12 EST"
[1] "Fitting WILMINK...END"
[1] "2023-03-09 11:25:31 EST"
[1] "Output files..."
[1] "2023-03-09 11:25:31 EST"
[1] "Output files...END"
[1] "2023-03-09 11:25:42 EST"
[1] "Output plots..."
[1] "2023-03-09 11:25:42 EST"
[1] "plot_1001_8.png"
[1] "plot_1002_1.png"
[1] "plot_1024_1.png"
[1] "plot_1025_4.png"
[1] "plot_1035_2.png"
[1] "plot_1039_7.png"
[1] "plot_1048_1.png"
[1] "plot_1063_1.png"
[1] "plot_1077_3.png"
[1] "plot_1104_6.png"
[1] "plot_1112_1.png"
[1] "plot_1117_1.png"
[1] "plot_1135_6.png"
[1] "plot_1138_5.png"
[1] "plot_1141_4.png"
[1] "plot_1152_5.png"
[1] "plot_1181_3.png"
[1] "plot_1183_4.png"
[1] "plot_1206_3.png"
[1] "plot_1215_8.png"
[1] "plot_1300_2.png"
[1] "plot_1313_5.png"
[1] "plot_1316_3.png"
[1] "plot_1346_2.png"
[1] "plot_1369_1.png"
[1] "plot_1375_2.png"
[1] "plot_1376_7.png"
[1] "plot_1389_5.png"
[1] "plot_1404_1.png"
[1] "plot_1407_1.png"
[1] "plot_1414_2.png"
[1] "plot_1416_3.png"
[1] "plot_1422_1.png"
[1] "plot_1428_3.png"
[1] "plot_1429_2.png"
[1] "plot_1450_6.png"
[1] "plot_1456_1.png"
[1] "plot_1459_1.png"
[1] "plot_1495_1.png"
[1] "plot_1509_1.png"
[1] "plot_1512_1.png"
[1] "plot_1514_3.png"
[1] "plot_1517_1.png"
[1] "plot_1523_1.png"
[1] "plot_1525_4.png"
[1] "plot_1528_3.png"
[1] "plot_1529_3.png"
[1] "plot_1530_1.png"
[1] "plot_1560_1.png"
[1] "plot_1581_2.png"
[1] "plot_1603_3.png"
[1] "plot_1607_2.png"
[1] "plot_1642_1.png"
[1] "plot_1665_5.png"
[1] "plot_1669_2.png"
[1] "plot_1670_2.png"
[1] "plot_1687_1.png"
[1] "plot_1703_1.png"
[1] "plot_1705_3.png"
[1] "plot_1719_5.png"
[1] "plot_1722_3.png"
[1] "plot_1738_3.png"
[1] "plot_1767_5.png"
[1] "plot_1776_4.png"
[1] "plot_1789_2.png"
[1] "plot_1797_2.png"
[1] "plot_1807_2.png"
[1] "plot_1836_1.png"
[1] "plot_1854_3.png"
[1] "plot_1856_7.png"
[1] "plot_1876_4.png"
[1] "plot_1877_5.png"
[1] "plot_1898_4.png"
[1] "plot_1917_1.png"
[1] "plot_1925_2.png"
[1] "plot_1950_3.png"
[1] "plot_1966_2.png"
[1] "plot_1970_3.png"
[1] "plot_1971_3.png"
[1] "plot_1975_1.png"
[1] "plot_1978_2.png"
[1] "Output plots...END"
[1] "2023-03-09 11:31:28 EST"
[1] "END"
```
