==> 4.Forecast_iteration_example.Rmd <==
---
title: "4. Forecast and iteration example"
output: 
  html_notebook: 
    fig_width: 8
    toc: yes
    code_folding: show
author: "David Innes"
editor_options: 
  chunk_output_type: inline
---

Like step 2, this is best executed on a larger computer. The script `4.Forecast_Dijkstra_parallel.R` is a wrapper around `f_iterate_fitting.R` which actually fits each level of DIM_threshold. 

It also requires user input to determining if you are running with 'constrained' or 'unconstrained' parameters. See script to set these values, should be based on step 4a. Scripts are executed in following steps with `is_constrained=FALSE`.

Takes the filtered_cleaned_output file from step 4a as input, but confirms with user input. 

Requires:
```{r eval=FALSE}

suppressPackageStartupMessages(library(tidyverse, warn.conflicts = FALSE))
library(nlme)
library(nls.multstart)
library(broom)
library(purrr)
library(multidplyr)
# requires car
source("./functions/f_lac_equations.R")
source("./functions/f_goodness_of_fit.R")
source("./functions/f_iterate_fitting.R")
source("./functions/f_forecast_P.R")

```

# Execute
This can also be slow, even with model fitting split across all available cores (-1). Output files are automatically saved to `Outputs/`. Looks for `filtered_cleaned_output____.rds` in this folder as well.
Keep an eye on console for user input.

```{r}
source('./4.Forecast_Dijkstra_parallel.R')
```

output:

```{}
[1] "Fitting with `is_constrained` ==  FALSE"
[1] "Starting DIM_threshold = 30"
[1] "2023-03-09 12:28:42 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:28:46 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:30:22 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 60"
[1] "2023-03-09 12:30:24 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:30:28 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:32:15 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 90"
[1] "2023-03-09 12:32:17 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:32:21 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:34:09 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 120"
[1] "2023-03-09 12:34:11 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:34:16 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:36:01 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 150"
[1] "2023-03-09 12:36:04 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:36:09 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:38:16 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 200"
[1] "2023-03-09 12:38:18 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:38:23 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:40:26 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 250"
[1] "2023-03-09 12:40:28 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:40:33 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:42:53 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Starting DIM_threshold = 300"
[1] "2023-03-09 12:42:55 EST"
[1] "Preparing Data..."
[1] "Using bounds on b param: 1e-09 to 1"
[1] "Using bounds on b0 param: 1e-09 to 1"
[1] "Using bounds on c param: 1e-09 to 1"
[1] "Preparing Data...END"
[1] "Fitting DIJKSTRA..."
[1] "2023-03-09 12:43:01 EST"
[1] "Fitting DIJKSTRA... END"
[1] "2023-03-09 12:45:06 EST"
[1] "Calculting P values..."
[1] "Calculting P values...END"
[1] "Fitting models complete"
[1] "Combining results..."
[1] "Combining results...END"
[1] "Exporting results..."
[1] "Exporting results...END"
[1] "2023-03-09 12:45:20 EST"
[1] "END"
```
